The organization of the project is restructured and the version of Material UI is updated
